fix: refactorization folder business
There were a big amount of scripts in the user folder so I had to move them all to another called business
