+++ src/main/java/com/streetask/app/business/BusinessAccountRepository.java
@@ -1,4 +1,4 @@
-package com.streetask.app.user;
+package com.streetask.app.business;
 
 import java.util.UUID;
 import java.util.Optional;
