+++ src/main/java/com/streetask/app/business/BusinessSubscriptionRestController.java
@@ -1,4 +1,4 @@
-package com.streetask.app.user;
+package com.streetask.app.business;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
