+++ src/main/java/com/streetask/app/business/BusinessSubscriptionStatusResponse.java
@@ -1,4 +1,4 @@
-package com.streetask.app.user;
+package com.streetask.app.business;
 
 import java.time.LocalDateTime;
 import java.util.UUID;
