+++ src/main/java/com/streetask/app/business/PublicStripeCheckoutSessionConfirmRequest.java
@@ -1,4 +1,4 @@
-package com.streetask.app.user;
+package com.streetask.app.business;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
