+++ src/main/java/com/streetask/app/business/RequestStatus.java
@@ -1,4 +1,4 @@
-package com.streetask.app.user;
+package com.streetask.app.business;
 
 public enum RequestStatus {
     PENDING,
